Moved to AccumulateOnlyPatternTest
