+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/AccumulateOnlyPatternTest.java
@@ -24,6 +24,7 @@ import java.util.Collection;
 import java.util.GregorianCalendar;
 import java.util.List;
 
+import org.apache.commons.math3.util.Pair;
 import org.assertj.core.api.Assertions;
 import org.drools.modelcompiler.domain.Person;
 import org.drools.modelcompiler.domain.Result;
@@ -179,5 +180,46 @@ public class AccumulateOnlyPatternTest extends OnlyPatternTest {
         Collection<Result> results = getObjectsIntoList(ksession, Result.class);
         assertEquals(1, results.size());
         assertEquals(2L, results.iterator().next().getValue());
+    }
+
+    @Test
+    public void testAccumulateWithIndirectArgument() {
+        String str =
+                "global java.util.List resultTotal; \n" +
+                        "global java.util.List resultPair; \n" +
+                        "import " + Person.class.getCanonicalName() + ";\n" +
+                        "import " + Pair.class.getCanonicalName() + ";\n" +
+                        "rule R " +
+                        "    when\n" +
+                        "        accumulate(\n" +
+                        "            Person($n1 : name)\n" +
+                        "            and Person($n2 : name);\n" +
+                        "            $pair :  collectList(Pair.create($n1, $n2)), \n" +
+                        "            $total : count(Pair.create($n1, $n2))\n" +
+                        "        )\n" +
+                        "    then\n" +
+                        "        resultTotal.add($total);\n" +
+                        "        resultPair.add($pair);\n" +
+                        "end";
+
+        KieSession ksession = getKieSession( str );
+
+        final List<Number> resultTotal = new ArrayList<>();
+        ksession.setGlobal("resultTotal", resultTotal);
+
+        final List<List<Pair>> resultPair = new ArrayList<>();
+        ksession.setGlobal("resultPair", resultPair);
+
+        Person mario = new Person("Mario", 40);
+        ksession.insert(mario);
+
+        int rulesFired = ksession.fireAllRules();
+        Assertions.assertThat(rulesFired).isGreaterThan(0);
+        Assertions.assertThat(resultTotal).contains(1l);
+
+        List<Pair> resultPairItem = resultPair.iterator().next();
+        Pair firstPair = resultPairItem.iterator().next();
+        assertEquals("Mario", firstPair.getFirst());
+        assertEquals("Mario", firstPair.getSecond());
     }
 }
