+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/AccumulateTest.java
@@ -2551,74 +2551,4 @@ public class AccumulateTest extends BaseModelTest {
             this.employee = employee;
         }
     }
-
-    @Test
-    public void testAccumulateWithIndirectArgument() {
-        String str =
-                "global java.util.List resultTotal; \n" +
-                "global java.util.List resultPair; \n" +
-                "import " + Person.class.getCanonicalName() + ";\n" +
-                "import " + Pair.class.getCanonicalName() + ";\n" +
-                "rule R " +
-                "    when\n" +
-                "        accumulate(\n" +
-                "            Person($n1 : name)\n" +
-                "            and Person($n2 : name);\n" +
-                "            $pair :  collectList(Pair.create($n1, $n2)), \n" +
-                "            $total : count(Pair.create($n1, $n2))\n" +
-                "        )\n" +
-                "    then\n" +
-                "        resultTotal.add($total);\n" +
-                "        resultPair.add($pair);\n" +
-                "end";
-
-        KieSession ksession = getKieSession( str );
-
-        final List<Number> resultTotal = new ArrayList<>();
-        ksession.setGlobal("resultTotal", resultTotal);
-
-        final List<List<Pair>> resultPair = new ArrayList<>();
-        ksession.setGlobal("resultPair", resultPair);
-
-        Person mario = new Person("Mario", 40);
-        ksession.insert(mario);
-
-        int rulesFired = ksession.fireAllRules();
-        Assertions.assertThat(rulesFired).isGreaterThan(0);
-        Assertions.assertThat(resultTotal).contains(1l);
-
-        List<Pair> resultPairItem = resultPair.iterator().next();
-        Pair firstPair = resultPairItem.iterator().next();
-        assertEquals("Mario", firstPair.getFirst());
-        assertEquals("Mario", firstPair.getSecond());
-    }
-
-    @Test
-    public void testAccumulateOnTwoPatterns2() {
-        // DROOLS-5738
-        String str =
-                "global java.util.List resultTotal; \n" +
-                        "import " + Person.class.getCanonicalName() + ";\n" +
-                        "import " + Pair.class.getCanonicalName() + ";\n" +
-                        "rule R1 when\n" +
-                        "  accumulate (\n" +
-                        "       $p1: Person( $a1 : age) and $p2: Person( $a2 : age ); $sum : collectList(Pair.create($a1, $a2)) " +
-                        "         )" +
-                        "then\n" +
-                        "  resultTotal.add($sum);\n" +
-                        "end\n";
-
-        KieSession ksession = getKieSession( str );
-
-        final List<List<Pair>> resultTotal = new ArrayList<>();
-        ksession.setGlobal("resultTotal", resultTotal);
-
-        ksession.insert( new Person( "Mario", 46 ) );
-
-        ksession.fireAllRules();
-
-        List<List<Pair>> results = resultTotal;
-        assertEquals( 1, results.size() );
-        assertEquals( 46, results.get(0).get(0).getFirst() );
-    }
 }
